Adds Facts table for metrics
The model `Facts::Metric` represents a fact with performance metrics
about a content item in a moment in time.

The grain is the Content Item and the frequency is daily. For this
reason, for each fact, we associate the following foreign_keys:

- Date: represents when the meassure has been taken
- Oragnisation: publishing organisation of the content item
- Item: content item with its attributes

This approach continues the star schema approach of a Data Warehouse.