Validate uniqueness of Dimensions::Organisation
Ensure that each instance of Dimensions::Organisation is unique
by scoping uniqueness of one attribute against all the others. If all
of the attributes are the same as another instance of a
Dimensions::Organisation then the instance will not be considered
unique. If the combination of attributes is unique then the instance
will be considered unique.