Improve performance of etl::items
We set up a 'dimensions_items_temps' table to act as a temporary
item store so that we could find the items that needed to be added to
the 'dimensions_items' table quickly and easily by comparing the two
tables. This reduced the time, and the number of database queries,
needed to ETL the items.

We also added a `latest` field to Dimensions::Items so that we can
easily identify, and return, the latest version of each item.