Add ADR to ignore translations to other languajes
Some Content Items have multiple translation of the content, so the
[Publishing-api][1] will return the `content_id` along with all locales
assigned to the Content Item.

We have decided to ignore content metrics and Analytics metrics for
non-english content for the first iterations.

The main reason is that we would need different algorithms and libraries
to make our application consistent among all the languages/ locales.
If this is a real need, then we will support it in future iterations of
the Data Warehouse.

The bigger benefit is that it makes the codebase much simpler as the
grain in the Fact Table for metrics linked to the Content Item's
dimension.

[1]: http://github.com/alphagov/publishing-api