Add an ETL source for Google Analytics report data
This will fetch pageview and unique pageview metrics by page path
for a given day from the Google Analytics Reporting API.

The data is returned as an enumerable collection so that it can be
traversed, searched, filtered etc as part of a pipelines of steps.

The data returned is not filtered so some values may need to be
discarded in a future pipeline step. This is the optimal way to
use the Reporting API as it produces a  computationally inexpensive
query.

The API client will automatically page through the response data
returned by the API. We fetch 10,000 rows of data in each query to
the API, this is the maximum supported and ensures we use our daily
quota of API requests economically.