Add an ETL transform for Google Analytics report data
This takes a collection of Google Analytics report data and filters
out any rows where we don't have a content item dimension
matching the page path.

Only a subset of the GOV.UK content is viewed everyday by someone,
so the collection will be smaller in size than the total number of
know content item dimensions.

The data returned is transformed with the appropriate values so that
it can be imported into the fact table.