Store Google Analytics events for processing
We have changed the approach of pulling Google Analytics events. We now
go through all the visits for a particular day and fetch GA metrics for
each path.

The Content Items base_path sometimes does not match the path on GA.
This happens with Guides, for example, so we won’t be able to find
directly a match between a Content Item and an Event from GA.

For this reason, I have decided to store the list of events in the
database so we can do:

1. A quick update of the content items with a matching path
2. A batch processing of the links that do not match.