Load GA events in Facts Table
I have split the process in two steps. The first one will import all
the events from GA in batches of 10_000 events. I have chosen this number
because it matches the maximum number of events in each GA request.
All the events are stored in GA::Events table to be process in the second
step.

On the second step, in a single query, we update the facts metrics table.
I have chosen to write custom SQL as the amount of records that we are
handling is very big (around 600K), and loading into AR for processing,
is taking almost 2 hours with a single process. The time is not a problem
as we don’t have performance requirements so far, but we should’t have
transactions that take so much time.

By running the update query we reduce this process to 7-8 seconds, so we
have a small transaction, and the database does all the heavy lifting,
not the application.