Update the Content Items Dimension On Publishing-api Events.
This is needed for the Update the Content Items dimension on
publishing-api events card. Set dirty flag in Dimensions::Item on
publishing api event when the item exists. Create a new
Dimensions::Item with dirty and latest flags set to true if it does
not exist.

Duplicate content items dimension and fire Sidekiq job for items marked
as dirty when ETL::Master#process is run.

Does not handle unpublish events.