Remove ActiveRecord validation for GA metrics
There is no need to validate GA metrics at this point. This validation
should be done, if we need to, in the ETL process.

In any case, we have not found so far any situation where this has
happened, and if this is the case, we are not going to take any immediate
action; as this is going to be an exception that do not affect the
overall ETL process, I am not adding this validation in the ETL process
for the time being.