Add documentation skeleton
This uses the GDS tech docs template and an OpenAPI spec that is converted to a markdown template using widdershins.

The config and markdown source will be added in a separate commit.

This code was forked from the Content API documentation: https://github.com/alphagov/govuk-content-api-docs

If we're happy with this pattern, we might want to think about making this
more reusable (perhaps integrating with https://github.com/alphagov/govuk-rails-app-template)