Upgrade to latest tech docs template
This PR upgrades the tech docs to the latest version, which uses the
[govuk_tech_docs gem](https://github.com/alphagov/tech-docs-gem) to
pull in CSS, JS, fonts, images and layouts. This means that a lot of
files can be removed and only customised code and configuration
remains. You can read more about it in
https://github.com/alphagov/tech-docs-template/pull/144.

A section on keeping the template up to date has been added to the
README.