Extract Job to run quality metrics
We have a different source to retrieve quality metrics, so it makes sense
to to the processing in its own job.

If, for any, reason the service that provides quality metrics is not
working we will still retrieve the JSON, and once the service is up again
we will populate the quality metrics.