Split metrics show endpoint into time series and summary
I think the two parts of the current response correspond to different use
cases.

One is useful for graphing things and extracting raw data, the other is useful
for showing aggregate values.

When viewing a content item you might want to view both of these kinds of
information, but I think the API will be easier to understand if we separate
the two, so I've created two new endpoints.

Note than when aggregating values over time, it's not always appropriate to sum
the values. Sometimes you'll want to average the values instead, or show the
latest value. This depends on whether the metric is counting events, or if it's
a snapshot of something.

The reference documentation will be expanded later to describe the response
structure and explain which metrics are available.