Add response definitions to the API reference
This generates example responses for each endpoint.

It also adds an object reference at the bottom of the page that shows
what each field means. This doesn't work very well at the moment
because widdershins renders "additionalProperties" as if it is a
literal field called "additionalProperties". I'll see if we
can fix this separately.