Change publishing api consumer to store every edition
Create a new edition every time we receive a major/minor/unpublish
event, and make that item the only version with latest=true.

During the day, the outdated flag is set on anything processed by
the stream, and remains set on every edition until it's processed
by the OutdatedItemsProcessor.

I'm explicitly only storing the content_id, base_path, status and locale
from the consumer, because everything else is currently populated
overnight. I plan to make the consumer solely responsible for populating
the items in a later commit.