Make OutdatedItemsProcessor process outdated items
OutdatedItemProcessor now processes anything that has the outdated flag
set to true. It sets the flag to false once the content details have
been processed.

This decouples the concept of "outdatedness" (which I want to get rid of)
from the concept of "latest". Only "latest" versions will have daily metrics,
but the latest flag is set in the publishing API consumer.