Only process messages in increasing version order
Ignore any messages that have a "payload_version" the same as
or less than that of the last message we received for a content item.
This number comes from the publishing event log, so any time we
observe numbers going down, we know we're processing things
in the wrong order.

RabbitMQ guarantees at-least-once delivery when using acknowledgements,
so messages can be duplicated, plus we can't guarantee the order of
processing because we can have multiple consumers working in parallel.

Usually this is not a problem, but on rare occasions we might
process multiple messages for the same content item in a short time.

The important thing for this story is to ensure that our "latest" record
is definitely the latest. Otherwise out of date metrics will get into
the fact table.

For simplicity I've just dropped any messages that are older than
the current latest version.

This logic could later be extended to also process out of date messages,
as long as we preserve the latest flag on the existing item, but I
don't think we need it right now.