Store edition level facts in the new facts table
Currently we store these on the items table as well as a separate facts table.
We'll remove the item columns in a future commit.

Our concept of version is slightly different from the publishing API's.
We consider something a new version if its JSON representation
has changed. At the moment Editions::Facts will respect this but
Dimensions::Item records are created for every major/minor edition.

To avoid storing redundant items we'd need to check the content hash in the
message processor, and only import content details if we need it.

I'm not doing this now because

1. The message queue JSON is different from the content store one, so this will
   invalidate all our stored hashes
2. It's possible the value we're hashing includes data that would vary even
   for versions we'd want to ignore (like timestamps)