Ignore duplicated items
When quering the PublishingAPi for all the Content Items in `published`
state, we are receiving multiple elements with the same information. We
have noticed this problem after checking that our Google Analytics
metrics were not accurate.

The root problem is that we are having multiple base_path `/` with the
attribute `latest=true` in the Items dimension. Interestingly all the
items were created at the same time, so it seems that the only point
where this could happen is in the Preload phase.

We are going to add a Database constraint to prevent having multiple
items with the `latest=true` attribute for the same basepath, but we
also need to prevent this problem from happening when querying for data.