Add base_path to InvalidSchemaError
This commit is part of a story to 'Support all schemas when parsing
content items'. Currently content schemas wer are not extracting
are being recorded as InvalidSchemaErrors.

We need to make a decision as to whether we want to store the
content from these 'Invalid' content schemas or not. Storing the
base_path will help us make decisions as we can access the specific
content relating to the error message.