Use partial unique index for unique base_paths
We could have multiple base_paths when the `latest` attribute is false, 
but only one when `latest` is true.

This index guarantees that our database is consistent with the way we
model the data. I have decided to add a constraint in the database as 
the performance of the operations is much faster, and it is likely that
other processes that might not depend on AR, perform operations on the 
database creating situations with inaccurate data that are not easy to 
detect.

For example, we could insert `pageviews` to the same `base_path` multiple
times making all calculations to be inaccurate.