Update sandbox to use Facts::Edition
This is a large commit as it took a lot of work to get to a green
state with the tests unfortunately. There were a lot of changes
to achieve a passing build.

But now that we have a set of passing tests, we can start to
refactor with smaller commits.