Fix `1 unique_pageviews` bug
We have a bug in our codebase which shows unique_pageviews for some pages to be 1 occasionally, in our
facts_metrics table, despite the true number usually being in the thousands.

The cause of this issue is that we expect to only have one GA event per page path which in turn feeds in
to one facts_metric per base path. However, for some pages, GA returns us data for the base path (i.e '/tax')
AND returns us data for the full url (i.e 'https://gov.uk/tax').

In our codebase we had been sanitising the events that contained the full url, transforming it into the base_path
that our codebase requires. However, when this happens we sometimes have 2 GA events in our table with the same
page path (one which had the basepath from GA, the other which had the full URL but we sanatised and which then was
identical to the basepath of the original event). This meant that as we loaded the data from events_gas into
facts_metrics the metrics from the event we loaded second would overwrite the metrics from the event that we loaded
first, resulting in some unexpected metrics.

In this PR I have made changes to the logic to the transform stage of our GA ETL process so that after sanitising any
page_paths that do not match our required format we also check that the newly sanatised page_path does not duplicate
the page_path of existing events. If it does then we update our event with the sum of the metrics of both events and
we delete the duplicate event. This will fix the problem.

I have also written a test to check this functionality and changed some names to reflect the new functionality.