Define metric type in the yaml file
It is better for the Metric class to know little about each one of the
metrics. I am planning to move the metric type into the metric definition
but for the time being, having the metrics grouped by type in the YML
file improve the readability and maintainability of the codebase.