Default events_gas pageviews and unique_pageviews columns to 0
A bug exists in our codebase which raises errors when we import GA metrics that do not
include `pageviews` or `unique_pageviews` in their responses. This is because we attempt
to aggregate these metrics when we check for duplication. When we don't receive either of
these metrics in the response from GA our database stores them as `nil` then the code
attempts to add `nil` to an integer and we get an error.

In this PR we change these metrics to default to 0 to fix this bug.