Add RSpec tests for RabbitMQ rake tasks
We need to test the rake tasks because they are key for the normal
operation of the application. They are not a one-off thing, nor even
a single task that is run by a user.

These rake tasks are run automatically and, unless we write tests for
them we could have an invalid version of the application running in
production.