Remove Content Store dependency
Removing the Content Store has a big impact in the codebase because
it affects to how the metadata is being calculated and all the integration
tests.

It also impacts the message processor, and how we parse the information
based no base_paths and not in content ids. For this reason, I am removing
all the code related to:

- Metadata that depends of Content Store
- Content Store
- Tests that setup the Content Store

In this PR I will test drive our new approach to event processing to
grow the items dimension.

There are four pending tests that we need to move into a green state before
preparing the PR.