Remove default locale to english
It is not a good idea to enforce a locale when the data does not exist
in PublishingAPI. I think this was added when the primary key for the
dimension was the locale plus the content_id.

This is no longer the case, so it is better to remove it.