Schema 'role' does not exist
The schema_name 'role' is a newly created Schema. At this stage,
it has been assessed that there is limited value in storing the
content for this schema. Therefore is has been added to the the
list of schemas that we are not extracting content from.

This will eliminate any InvalidSchemaError messages for this schema.