Extract the individual parts of multi-part pages
Previously we used to smoosh together all the parts of a multipart
document. Now we want to store them separately, so just extract the
body text.

Note that `body` can sometimes be an array instead of a string.
This is the same issue fixed in the body parser here:
https://github.com/alphagov/content-performance-manager/pull/756

Because ruby's join method silently converts non-string objects
to strings, this was not raising an error before, but we were
mangling the text we were passing to the quality metric service.

This commit also moves the content parser call to the processor,
because the model shouldn't need to know about this. It's part
of the processing.