Fix undefined method 'delete!' for Array in Licence schema
Licence parser can now handle if the Json 'licence_overview'
field is a String or an Array. If it is an array we only want
the element that has the content_type of "text/html".

We explicitly return nil if the 'licence_overview' field is an
empty string or an empty array. This will eliminate the errors
that we were receiving.