Tidy up tests for Message Adapter
Some of the tests don't quite fit into a unit test. Tidying up so that
refactoring out later can be a bit easier. Also the schemas allow
multi part content to have a type of both `text/govspeak` and
`text/html` so we should have test data that contains both.