Fix: Invalid namespace when populating metrics
There was an error in the etl::master process.

The error was not caught by tests because the method with the error
was stubbed out in the tests.

The test is now stubbed out at a lower level, and the bug is caught.

Code fixed and should not be able to happen again.