Quality metrics service uses simpler response
The quality metrics (node) service has been experiencing
a 'Payload Too Large' error when sending a large block of text.

This is due to the service returning the details of every error
found.

The service has now been refactored to return the following format:
{
  readability_count: 1,
  contractions_count: 2,
  equality_count: 3,
}

without any of the detail.

This commit changes the local client of this service to use the
new response format.