Add PartsAdapter class
It's a bad name but can rebase later.

This class extracts parts out of a JSON message from the Publishing API,
for the MessageAdapter class to send to the Parsers. The MessageAdapter
sends each part as is they are single content items, which is how we
store them as in the Data Warehouse, as each has its own URL.