Prevent edge errors when processing in parallel
When processing messages in pararell we could have situations where two
updates for the same content item as being processed at the same time.
This is a very unlikely scenario but it is actually hapenning with big
republishings.

We have a constraint in the Database that ensures that for a given
basepath we only have one item with the attribute latest set to true.

If two consumers are processing an update for the same content item they
would be trying to update the latest attribute to true at the same time
and the constraint will prevent one of the updates from happening. In
that moment the following exception is being raised: ActiveRecord::RecordNotUnique

If this happens it is ok to just retry the operation a few times,
because eventually the message will be processed successfully.