Remove unique index on base_path and latest
We are having many errors and unexpected behaviours and we are not sure
if this constraint is preventing the normal flow of our ETL processes.

We plan to restore it with an AR validation, but for the time being we
prefer to delete it so we can be more confident about the root cause of
our problems.