Discard old messages before processing
We need to discard the messages when they are old because we can’t take
any action based on old data.

We were having issues because old messages were deprecating all our
content items, which is something not exposed in our tests; I have added
an integration test so this situation don’t happen again.