Ensure a link update don't deprecate all items
When we receive a link update and the update contains no changes then we
deprecate all the content items which implies that we won't be storing
the latest version of that content item.

The way we currently process messages is not very reasy to read and side
effects can happen without a clear representation of the cause in the
codebase. This is something that we will deal with very soon. In the
meantime, I have covered this logic with an integration test.