Reduce number of quality metrics
Our team will only use word_count in the first iteration of this app which means
that we have no need to make calls to the govuk-content-quality-service.

In this PR we disable requests to the above service and calculate the only
quality metric we currently require for our mvp, word count, within our own codebase.

We continue to calculate the other metrics that we can get without making a
call to the quality metrics service as they basically come for free.

Note: Performing the population of quality metrics requires us to validate that
new_items have document_text. This was not previously needed as only the
job concerned with the item with no document text would fail when run async.