Use govuk-app-config healthcheck support
We have implemented our own way of supporting healthchecks, which is ok
but we should be using the gem that GOV.UK has created for all the
applications.

This update is also related to an incidence when the existing
implementation of the healthcheck was causing issues with a memory leak
with Rails 5.2.1

The existing code is the only running code in Staging at interval of
times, so we think that the existing code for the healthcheck might have
some memory leak with Rails 5.2.1

In any case, we should use the gem support for providing a healthcheck.