Extract Item promotion object
Promotions involved a change check plus an action if the promotion
should happen. To validate the change we need to be sure that there is
a change in any of the attributes that make the new item different
from the previous one.

For this reason I have extracted an object that at this moment only
validates if an Item has changed, but that will also perform the
promotion later on.

This is an intermediate step in a longer refactor that aims to simplify
the way we process publishing-api events.