Fix ETL not calculating satisfaction_score
The ETL process was not calculating the satisfaction_score within the
SQL that we execute in order to cope with setting a huge number of
updates. Previously we were only using the the ActiveModel overrides of
the setters for `is_this_useful_yes` and `is_this_useful_no` to then
calculate and set the value of `satisfaction_score` but the SQL bypassed
that mechanism.