Add content_uuid to dimensions_items
We need a way to identify the same item of content.
Until now we have been using base_path to identify an
individual piece of content.

If the base path changes, we will lose the link between the
editions. All metrics collected from GA will be linked to the
individual editions.

The content_uuid will be populated by the Etl process and
will link different versions together.

We can then use this coloumn to join the latest item to
all previous editions, and summarise the metrics across all
editions.