ADR: add unique identifier to the DW grain
ADR to associate a unique identifier to each item that is tracked in the
Items Dimension.

Each `base_path`, which represents the grain of Data Warehouse, will
have a unique identifier that will never change. This ID will allow us
to perform queries through time on a efficient way, and also to be
accurate with our results.