Extract TimeSeriesController from MetricsController
Metrics controller is doing too many things; this might be because it
has a very generic name, which is a smell in itself.

I have decided to split MetricsController into smaller chunks so they
can do a single thing and have more descriptive names.

The first controller that I have extracted is TimeSeriesController with
a show action on it.