Cut down the parts of Rails that are loaded
Some parts of Rails loaded by default are not being used, and probably
won't ever be, like actionmailer for example. So just require the
parts that are used.

The motivation behind this change is to remove activestorage from
being required, as it requires secrets when precompiling assets, which
breaks it in the GOV.UK Mini Environment Admin.