Remove duplicated editions with same base_path
We should only have an edition in its latest state (latest = true) per
base_path.

In the script we first identify the id of editions that have more than
one `base_path` in their latest edition. Then we iterate through
them and destroy the facts_edition and the edition.

This is a process that need to be run until no duplications are found.