Retries Job before raising Sentry errors
We expect to have errors related to duplicated editions from time to time
because we can’t guarantee the order in which we receive the messages.
Sometimes they might run at the same time.

By adding the retry at the Job level, we aim to prevent having actionable
errors in Sentry.