Increase `work_mem` when running migrations
The `work_mem` parameters is used by PostgreSQL to determine how much
memory will reserve when running a query.

In our case, we need to set it to a high value because our views are
big aggregations that need to precalculate all values.

Obviously we can’t set this value for all the queries because we would
impact in the memory reserved for the Database, so we set it up only
for the current transaction by using `local`.