Use DateRange in FindContent query
There is no need to use `from` and `to` because all the views are
precalculated. So I am removing all references to `from` and `to` and
replacing them with `date_range`

I have not updated yet all the views that need to be used depending
on the date range. I am planning in doing it in the next commits