Improve SingleItemHandler for messages with no locale
Some messages from the Publishing API currently don't include a
locale, messages with a document_type of gone or redirect for example.

This commit changes the SingleItemHandler to not depend on the locale
for finding the old edition, falling back to the combination of the
content_id, base_path, latest attributes if the locale isn't contained
in the message.