Prevent unnecessary dimensions_editions growth for some items
Previously, when processing messages from the Publishing API without a
locale, the generated warehouse_message_id could differ from the
warehouse_item_id on the previous edition.

Due to the way that the warehouse_item_id is carried forward, this
happens over and over again, even if nothing is changing as the
warehouse_item_id that's used in the comparison isn't necessarily the
one that's saved.

To prevent this, exclude the warehouse_item_id when comparing the old
edition. Also add a test to test this scenario.

This should also be addressed by including the locale in the messages
for gone and redirect items for the Publishing API, but I think this
is still a good change.