Assign warehouse_item_id based on old edition
The `warehouse_item_id` is not known when the message is being built.
This happens in redirects because they can be applied to single and
multipart messages.

We have moved it to the handler because it is in this point when we
have a reference to the old edition, which allow us to get the
warehouse_item_id.