Show correct number of PDFs
We have a bug which shows an incorrect number of PDFs attachments for some content.

This is because we use regex to search for pdf but it is case sensitive.

This PR makes the regex expression case insensitive because some content has
attachments which have the pdf extension in upper case.