Add search by base_path to the Editions table
It uses PostgreSQL full text search capabilities to search though all
the base_paths. I have tried with pg_search[1], but I have finally
decided not to use it because there is no way to skip the sorting by
ranking.

In our scenario, we don't want the results to be sorted by ranking
because we already have an existing sorting criteria. It is just better
to rely on simple text searching with no aditional ranking.

I have created an index that significantly improves the performance of
the query.

Important to note here is that I am creating the index removing the `/`
because this was creating many issues are the `/` where considered letter
and making the whole match to fail.

Also important to note how the index and the query share the same way
to remove the `/`