Remove defaults from facts_metrics and events_gas
Why
When repopulating data we are getting zeros in metrics columns
across the dates we have repopulated. Some of these are from before
the item was published.

These columns should only have data when we have actually received it
from the source (i.e. it's been published and we collected some data).