Change the warehouse_item_ids for multipart messages
Previously the warehouse_item_ids for multipart messages included the
full path. This means that if the base_path for the content item in
the Publishing API were to change, the warehouse_item_id would be
different.

This somewhat defeats the point of the warehouse_item_id, which should
stay the same for the same page, even if it moves.

To improve the situation, just include the "sub" path, the bit after
the base path (from the Publishing API). For example, with a bit of
content with two pages /basepath and /basepath/subthing, the
warehouse_item_id of the two pages would be:

 - /basepath          => 176a6aab-...583e61f45:en
 - /basepath/subthing => 176a6aab-...583e61f45:en:subthing

As the warehouse_item_id can be used to spot uniqueness issues, even
though the dimensions_editions entry for /basepath corresponds to a
multipart message, having the warehouse_item_id not contain anything
special should be more effective at avoiding duplicates.

For the dimensions_editions entry for /basepath/subthing, I decided to
omit the leading /, as I though that might be confusing (as it would
look like an absolute path).