Allow /single_page route to handle the homepage.
Now we have allowed filtering by no organisation or
all organisations, the GOV.UK homepage now appears in the
results. This has highlighted a bug. The homepage's base_path
`/` doesn't map to a route in either app.

This commit makes the /*base_path section of the route optional
so when you call it with /single_page?date_range=past-30-days
it returns data for the homepage. Previously it would 404.