Fix total_count returned from /content
We have an error where the total_results returned from
this endpoint does not match the number of rows returned.

The cause of this is that `latest: true` is not included
in the criteria in `Queries::FindContent#slice_editions`
and `slice_editions.count` is how we retreive the `total_results`
returned.

Refactored the code to use `slice_editions.latest.count` and
added a test to check that old editions are excluded from the
count.