Fix incorrect pdf_count for mainstream documents
Mainstream documents are showing 0 for pdf_count.

The cause if this is that other content has the PDF links
under details > documents in the json and are nested under a div
with the class `attachment-details`.

Mainstream documents have the PDF links under details > body
and are nested under a div with the class `form-download`.

This commit changes the number_of_pdfs processor to find
pdf links in this location.